fix minimize.sh so it doesn't error if there are no swap partitions

+++ subrepos/image-creation/packer/scripts/minimize.sh
@@ -41,8 +41,10 @@ let count--
 dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count;
 rm /boot/whitespace;
 
-swappart=`cat /proc/swaps | tail -n1 | awk -F ' ' '{print $1}'`
-swapoff $swappart;
-dd if=/dev/zero of=$swappart;
-mkswap $swappart;
-swapon $swappart;
+swappart=$(cat /proc/swaps | awk 'BEGIN {FS=" "} NR>1 {print $1}')
+if [ -z "${swappart}" ]; then
+  swapoff $swappart;
+  dd if=/dev/zero of=$swappart;
+  mkswap $swappart;
+  swapon $swappart;
+fi
